Test coverage, code cleanup, formatting.
