+++ lib/src/cli/task_get_flutter.dart
@@ -18,6 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+import "dart:ffi";
 import "dart:io";
 
 import "package:meta/meta.dart";
@@ -101,6 +102,36 @@ class GetFlutterSDK extends Task<Directory> {
       _compatibleFlutterVersions[dist] ??
       (throw KlutterException(
           "Failed to determine download URL for Flutter SDK: ${dist.prettyPrintedString}"));
+}
+
+/// Find applicable [FlutterDistribution] for the current
+/// [OperatingSystem] and [Architecture] or throw [KlutterException].
+FlutterDistribution toFlutterDistributionOrThrow(
+    {required VerifiedFlutterVersion version,
+    required String pathToRoot,
+    PlatformWrapper? platformWrapper}) {
+  final p = platformWrapper ?? platform;
+  OperatingSystem? os;
+
+  if (version.os != null) {
+    os = version.os;
+  } else if (p.isWindows) {
+    os = OperatingSystem.windows;
+  } else if (p.isMacos) {
+    os = OperatingSystem.macos;
+  } else if (p.isLinux) {
+    os = OperatingSystem.linux;
+  } else {
+    throw KlutterException(
+        "Current OS is not supported (supported: macos, windows or linux): ${Platform.operatingSystem}");
+  }
+
+  final arch = version.arch ??
+      (Abi.current().toString().contains("arm")
+          ? Architecture.arm64
+          : Architecture.x64);
+
+  return FlutterDistribution(version: version.version, os: os!, arch: arch);
 }
 
 Map<FlutterDistribution, String> get _compatibleFlutterVersions {
