+++ pubspec.yaml
@@ -4,6 +4,9 @@ version: 3.0.0
 homepage: https://buijs.dev
 repository: https://github.com/buijs-dev/klutter-dart
 
+executables:
+  kradle:
+
 environment:
   sdk: ">=2.17.6 <4.0.0"
